requirements.txt
streamlit
pandas
plotly
python-dotenv
requests
databricks-sql-connector>=3.0.0